REF move _BufferedPartialReader and _retry to _files.py
